requirements.txt
psycopg2
pg8000

marshmallow-sqlalchemy

flask
flask-restful
flask-sqlalchemy
flask-marshmallow
flask-migrate
flask-shell-ipython
